requirements.txt
numpy
tensorflow
flask
pillow
